refs qorelanguage/qore#3171 split lang tests and jni tests into separate source files
